Move thread-related debug print functionality to threading_base module
