+++ hpx/runtime/threads/policies/queue_holder_thread.hpp
@@ -11,6 +11,7 @@
 #include <hpx/assertion.hpp>
 #include <hpx/debugging/print.hpp>
 #include <hpx/runtime/threads/policies/lockfree_queue_backends.hpp>
+#include <hpx/threading_base/print.hpp>
 #include <hpx/threading_base/thread_queue_init_parameters.hpp>
 #include <hpx/threading_base/thread_data.hpp>
 #include <hpx/threading_base/thread_data_stackful.hpp>
